Added support for Delete By Query API requests
Added a function for making requests to the Delete By Query API (Elastic
Search 5). As a side effect, the From and Size fields of the Search data
type were wrapped in Maybe, because they are not required for making a
search. This is best for searches anyway, let Elastic Search provide the
default, but it was required for Delete By Query API to work as the
requests were rejected if they were containing From/Size fields.

Types for handling the responses were not added as part of this patch.